COT - EMBED/requirements.txt
aiohttp
aiosignal
attrs
croniter
discord-components
discord-py-interactions
discord-typings
discord-ui
discord.py
emoji
frozenlist
idna
multidict
python-dateutil
python-dotenv
pytz
six
tomli
typing_extensions
yarl
audioop-lts